% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_names}
\alias{area_names}
\title{Banlgadesh administrative levels names in English}
\format{
A data frame with 5160 rows and 4 variables:

\describe{
\item{District}{district (admin level 2) names}
\item{Division}{division (admin level 1) names}
\item{Upazila}{upazila (admin level 3) names}
\item{Union}{upazila (admin level 3) names}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
area_names
}
\description{
A dataset containing Division, District, Upazila, and Union names
}
\keyword{datasets}
