% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_plot.R
\name{bd_plot}
\alias{bd_plot}
\title{sample function for plotting map of different administrative levels}
\usage{
bd_plot(level = "country", type = "static")
}
\arguments{
\item{level}{Administrative level of bangladesh. Should be one of: "country", "division", "district", "upazila","union'}

\item{type}{Plotting mode: "static" or "interactive"}
}
\value{
Static or interactive plot for administrative levels
}
\description{
uses tmap
}
\examples{
# Plot static map of district
bd_plot(level = "district", type = "static")


}
