% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_search.R
\name{bd_search}
\alias{bd_search}
\title{search for specific areas}
\usage{
bd_search(searchFor, level = "division", as.is = FALSE, coordinates = FALSE)
}
\arguments{
\item{searchFor}{search keyword}

\item{level}{administrative level of bangladesh. Should be one of: "country", "division", "district", "upazila","union'}

\item{as.is}{boolean, if TRUE, matches exact keyword as given}

\item{coordinates}{boolean, if TRUE, returns centroids of searched areas (latitudes and longitudes)}
}
\value{
A data frame
}
\description{
uses sf
}
\examples{
bd_search("amtali", level = "union", as.is = TRUE, coordinates = TRUE)
}
