% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{get centroids of administrative areas}
\usage{
get_coordinates(level = "division")
}
\arguments{
\item{level}{administrative level of bangladesh. Should be one of: "division", "district", "upazila","union'}
}
\value{
A data frame containing latitudes and longitudes
}
\description{
uses sf
}
\examples{
get_coordinates(level = "division")
get_coordinates(level = "district")

}
