% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divisions.R
\name{get_divisions}
\alias{get_divisions}
\title{get partial maps for divisions}
\usage{
get_divisions(divisions, level = "division")
}
\arguments{
\item{divisions}{character vector for division names. Can take multiple divisions.}

\item{level}{administrative level of bangladesh. Should be one of: "division", "district", "upazila","union'}
}
\value{
shapefile for given administrative level
}
\description{
get partial maps for divisions
}
\examples{
get_divisions(divisions = "Sylhet",level =  "upazila")
}
