% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_country}
\alias{map_country}
\title{Banlgadesh administrative level 0 shapefile}
\format{
A shapefile with 3 variables:

\describe{
\item{Country}{country (admin level 0) name}
\item{ADM0_PCODE}{admin level 0 codes}
\item{geometry}{MULTIPOLYGON for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
map_country
}
\description{
A shapefile containing level 0 administrative boundaries
}
\keyword{datasets}
