% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_district}
\alias{map_district}
\title{Banlgadesh administrative level 2 shapefile}
\format{
A shapefile with 7 variables:

\describe{
\item{District}{district (admin level 2) names}
\item{ADM2_PCODE}{admin level 2 codes}
\item{Division}{division (admin level 1) names}
\item{ADM1_PCODE}{admin level 1 codes}
\item{Country}{country (admin level 0) name}
\item{ADM0_PCODE}{admin level 0 codes}
\item{geometry}{MULTIPOLYGON for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
map_district
}
\description{
A shapefile containing level 2 administrative boundaries
}
\keyword{datasets}
