% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_division}
\alias{map_division}
\title{Banlgadesh administrative level 1 shapefile}
\format{
A shapefile with 5 variables:

\describe{
\item{Division}{division (admin level 1) names}
\item{ADM1_PCODE}{admin level 1 codes}
\item{Country}{country (admin level 0) name}
\item{ADM0_PCODE}{admin level 0 codes}
\item{geometry}{MULTIPOLYGON for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
map_division
}
\description{
A shapefile containing level 1 administrative boundaries
}
\keyword{datasets}
