% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_union}
\alias{map_union}
\title{Banlgadesh administrative level 4 shapefile}
\format{
A shapefile with 11 variables:

\describe{
\item{Union}{upazilla (admin level 4) names}
\item{ADM4_PCODE}{admin level 4 codes}
\item{Upazila}{upazilla (admin level 3) names}
\item{ADM3_PCODE}{admin level 3 codes}
\item{District}{district (admin level 2) names}
\item{ADM2_PCODE}{admin level 2 codes}
\item{Division}{division (admin level 1) names}
\item{ADM1_PCODE}{admin level 1 codes}
\item{Country}{country (admin level 0) name}
\item{ADM0_PCODE}{admin level 0 codes}
\item{geometry}{MULTIPOLYGON for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
map_union
}
\description{
A shapefile containing level 4 administrative boundaries
}
\keyword{datasets}
