% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_upazila}
\alias{map_upazila}
\title{Banlgadesh administrative level 3 shapefile}
\format{
A shapefile with 9 variables:

\describe{
\item{Upazila}{upazilla (admin level 3) names}
\item{ADM3_PCODE}{admin level 3 codes}
\item{District}{district (admin level 2) names}
\item{ADM2_PCODE}{admin level 2 codes}
\item{Division}{division (admin level 1) names}
\item{ADM1_PCODE}{admin level 1 codes}
\item{Country}{country (admin level 0) name}
\item{ADM0_PCODE}{admin level 0 codes}
\item{geometry}{MULTIPOLYGON for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
map_upazila
}
\description{
A shapefile containing level 3 administrative boundaries
}
\keyword{datasets}
