% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_district_2011}
\alias{pop_district_2011}
\title{Banlgadesh population census-2011 data for district level}
\format{
A data frame with 64 rows and 25 variables:

\describe{
\item{district}{district (admin level 2) names}
\item{admin2Pcode}{district codes}
\item{division}{division (admin level 1) names}
\item{admin1Pcode}{division codes}
\item{population}{population in 2011}
\item{P00_04}{population in age group 0-4}
\item{P05_09}{population in age group 5-9}
\item{P10_14}{population in age group 10-14}
\item{P15_19}{population in age group 15-19}
\item{P20_24}{population in age group 20-24}
\item{P25_29}{population in age group 25-29}
\item{P30_34}{population in age group 30-34}
\item{P35_39}{population in age group 35-39}
\item{P40_44}{population in age group 40-44}
\item{P45_49}{population in age group 45-49}
\item{P50_54}{population in age group 50-54}
\item{P55_59}{population in age group 55-59}
\item{P60_64}{population in age group 60-64}
\item{P65_69}{population in age group 65-69}
\item{P70_74}{population in age group 70-74}
\item{P75_80}{population in age group 75-80}
\item{P80plus}{population in age group 80+}
\item{Child}{child population}
\item{Male}{male population}
\item{Female}{female population}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
pop_district_2011
}
\description{
A dataset containing total population, population by age groups and gender
for each districts (administrative level 2) in bangladesh
}
\keyword{datasets}
