% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_plot.R
\name{bd_plot}
\alias{bd_plot}
\title{plot map}
\usage{
bd_plot(level = "country", type = "static")
}
\arguments{
\item{level}{administrative level of bangladesh. Should be one of: "country", "division", "district", "upazila","union'}

\item{type}{Plotting mode: "static" or "interactive"}
}
\description{
uses tmap
}
