% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_name.R
\name{get_area_names}
\alias{get_area_names}
\title{get area names in English, available in the shapefiles}
\usage{
get_area_names()
}
\value{
A data frame with area names in English
}
\description{
get area names in English, available in the shapefiles
}
\examples{
names <- get_area_names()

}
